edit: file upload update

+++ backend/controllers/noteController.js
@@ -1,5 +1,10 @@
 const { validationResult } = require("express-validator");
+
+// models
 const Note = require("../models/Note");
+
+// utils
+const { unlink } = require("../utils/unlink");
 
 exports.createNote = async (req, res) => {
   const { title, description } = req.body;
@@ -59,6 +64,11 @@ exports.getDetailNote = async (req, res) => {
 exports.deleteNote = async (req, res) => {
   const { id } = req.params;
   try {
+    const oldNote = await Note.findById(id);
+    if (!oldNote) {
+      return res.status(404).json({ message: "Note not found!" });
+    }
+    unlink(oldNote.cover_image);
     const note = await Note.findByIdAndDelete(id);
     if (!note) {
       return res.status(404).json({ message: "Note not found!" });
@@ -73,14 +83,20 @@ exports.deleteNote = async (req, res) => {
 exports.updateNote = async (req, res) => {
   const { id } = req.params;
   const { title, description, note_id } = req.body;
+  const cover_image = req.file;
 
   try {
     const note = await Note.findById(note_id);
+    console.log("note", note);
     if (!note) {
       return res.status(404).json({ message: "Note not found!" });
     }
     note.title = title;
     note.description = description;
+    if (cover_image) {
+      unlink(note.cover_image);
+      note.cover_image = cover_image.path;
+    }
     await note.save();
     res.status(200).json({ message: "Note updated!" });
   } catch (error) {
